fix: return type to `dto.ResponseSuccess`
