+++ src/app/service/pet/pet.service.go
@@ -8,9 +8,7 @@ import (
 	"github.com/isd-sgcu/johnjud-gateway/src/app/constant"
 	"github.com/isd-sgcu/johnjud-gateway/src/app/dto"
 	utils "github.com/isd-sgcu/johnjud-gateway/src/app/utils/pet"
-	"github.com/isd-sgcu/johnjud-gateway/src/constant/pet"
 	petproto "github.com/isd-sgcu/johnjud-go-proto/johnjud/backend/pet/v1"
-	imgproto "github.com/isd-sgcu/johnjud-go-proto/johnjud/file/image/v1"
 	"github.com/rs/zerolog/log"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -53,7 +51,7 @@ func (s *Service) FindAll() (result []*dto.PetResponse, err *dto.ResponseErr) {
 		}
 	}
 	imagesList := utils.MockImageList(len(res.Pets))
-	findAllResponse := RawToDtoList(res.Pets, imagesList)
+	findAllResponse := utils.RawToDtoList(res.Pets, imagesList)
 	return findAllResponse, nil
 }
 
@@ -92,7 +90,7 @@ func (s *Service) FindOne(id string) (result *dto.PetResponse, err *dto.Response
 		}
 	}
 	images := utils.MockImageList(1)[0]
-	findOneResponse := RawToDto(res.Pet, images)
+	findOneResponse := utils.RawToDto(res.Pet, images)
 	return findOneResponse, nil
 }
 
@@ -100,7 +98,7 @@ func (s *Service) Create(in *dto.CreatePetRequest) (ressult *dto.PetResponse, er
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
 
-	request := CreateDtoToRaw(in)
+	request := utils.CreateDtoToRaw(in)
 
 	res, errRes := s.petClient.Create(ctx, request)
 	if errRes != nil {
@@ -132,7 +130,7 @@ func (s *Service) Create(in *dto.CreatePetRequest) (ressult *dto.PetResponse, er
 		}
 	}
 	images := utils.MockImageList(1)[0]
-	createPetResponse := RawToDto(res.Pet, images)
+	createPetResponse := utils.RawToDto(res.Pet, images)
 	return createPetResponse, nil
 }
 
@@ -140,7 +138,7 @@ func (s *Service) Update(id string, in *dto.UpdatePetRequest) (result *dto.PetRe
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
 
-	request := UpdateDtoToRaw(id, in)
+	request := utils.UpdateDtoToRaw(id, in)
 
 	res, errRes := s.petClient.Update(ctx, request)
 	if errRes != nil {
@@ -178,7 +176,7 @@ func (s *Service) Update(id string, in *dto.UpdatePetRequest) (result *dto.PetRe
 		}
 	}
 	images := utils.MockImageList(1)[0]
-	updatePetResponse := RawToDto(res.Pet, images)
+	updatePetResponse := utils.RawToDto(res.Pet, images)
 	return updatePetResponse, nil
 }
 
@@ -272,138 +270,4 @@ func (s *Service) ChangeView(id string, in *dto.ChangeViewPetRequest) (result *d
 	return &dto.ChangeViewPetResponse{
 		Success: res.Success,
 	}, nil
-}
-
-func CreateDtoToRaw(in *dto.CreatePetRequest) *petproto.CreatePetRequest {
-	return &petproto.CreatePetRequest{
-		Pet: &petproto.Pet{
-			Type:         in.Type,
-			Species:      in.Species,
-			Name:         in.Name,
-			Birthdate:    in.Birthdate,
-			Gender:       petproto.Gender(in.Gender),
-			Habit:        in.Habit,
-			Caption:      in.Caption,
-			Images:       []*imgproto.Image{},
-			Status:       petproto.PetStatus(in.Status),
-			IsSterile:    *in.IsSterile,
-			IsVaccinated: *in.IsVaccinated,
-			IsVisible:    *in.IsVisible,
-			IsClubPet:    *in.IsClubPet,
-			Background:   in.Background,
-			Address:      in.Address,
-			Contact:      in.Contact,
-			AdoptBy:      in.AdoptBy,
-		},
-	}
-}
-
-func UpdateDtoToRaw(id string, in *dto.UpdatePetRequest) *petproto.UpdatePetRequest {
-	req := &petproto.UpdatePetRequest{
-		Pet: &petproto.Pet{
-			Id:         id,
-			Type:       in.Type,
-			Species:    in.Species,
-			Name:       in.Name,
-			Birthdate:  in.Birthdate,
-			Gender:     petproto.Gender(in.Gender),
-			Habit:      in.Habit,
-			Caption:    in.Caption,
-			Images:     []*imgproto.Image{},
-			Status:     petproto.PetStatus(in.Status),
-			Background: in.Background,
-			Address:    in.Address,
-			Contact:    in.Contact,
-			AdoptBy:    in.AdoptBy,
-		},
-	}
-
-	if in.IsClubPet == nil {
-		req.Pet.IsClubPet = false
-	} else {
-		req.Pet.IsClubPet = *in.IsClubPet
-	}
-
-	if in.IsSterile == nil {
-		req.Pet.IsSterile = false
-	} else {
-		req.Pet.IsSterile = *in.IsSterile
-	}
-
-	if in.IsVaccinated == nil {
-		req.Pet.IsVaccinated = false
-	} else {
-		req.Pet.IsVaccinated = *in.IsVaccinated
-	}
-
-	if in.IsVisible == nil {
-		req.Pet.IsVisible = false
-	} else {
-		req.Pet.IsVisible = *in.IsVisible
-	}
-
-	return req
-}
-
-func RawToDto(in *petproto.Pet, images []*imgproto.Image) *dto.PetResponse {
-	pet := &dto.PetResponse{
-		Id:           in.Id,
-		Type:         in.Type,
-		Species:      in.Species,
-		Name:         in.Name,
-		Birthdate:    in.Birthdate,
-		Gender:       pet.Gender(in.Gender),
-		Habit:        in.Habit,
-		Caption:      in.Caption,
-		Status:       pet.Status(in.Status),
-		IsSterile:    &in.IsSterile,
-		IsVaccinated: &in.IsVaccinated,
-		IsVisible:    &in.IsVisible,
-		IsClubPet:    &in.IsClubPet,
-		Background:   in.Background,
-		Address:      in.Address,
-		Contact:      in.Contact,
-		AdoptBy:      in.AdoptBy,
-		Images:       extractImages(images),
-	}
-	return pet
-}
-
-func RawToDtoList(in []*petproto.Pet, imagesList [][]*imgproto.Image) []*dto.PetResponse {
-	var resp []*dto.PetResponse
-	for i, p := range in {
-		pet := &dto.PetResponse{
-			Id:           p.Id,
-			Type:         p.Type,
-			Species:      p.Species,
-			Name:         p.Name,
-			Birthdate:    p.Birthdate,
-			Gender:       pet.Gender(p.Gender),
-			Habit:        p.Habit,
-			Caption:      p.Caption,
-			Status:       pet.Status(p.Status),
-			IsSterile:    &p.IsSterile,
-			IsVaccinated: &p.IsVaccinated,
-			IsVisible:    &p.IsVisible,
-			IsClubPet:    &p.IsClubPet,
-			Background:   p.Background,
-			Address:      p.Address,
-			Contact:      p.Contact,
-			AdoptBy:      p.AdoptBy,
-			Images:       extractImages(imagesList[i]),
-		}
-		resp = append(resp, pet)
-	}
-	return resp
-}
-
-func extractImages(images []*imgproto.Image) []dto.ImageResponse {
-	var result []dto.ImageResponse
-	for _, img := range images {
-		result = append(result, dto.ImageResponse{
-			Id:  img.Id,
-			Url: img.ImageUrl,
-		})
-	}
-	return result
 }
